+++ src/options.js
@@ -68,8 +68,13 @@ class RecrasOptions {
         if (!options.recras_hostname) {
             throw new Error(this.languageHelper.translate('ERR_NO_HOSTNAME'));
         }
-        if (!hostnameRegex.test(options.recras_hostname) && RecrasOptions.hostnamesDebug.includes(options.recras_hostname)) {
-            throw new Error(this.languageHelper.translate('ERR_INVALID_HOSTNAME'));
+        if (
+            !hostnameRegex.test(options.recras_hostname) &&
+            !RecrasOptions.hostnamesDebug.includes(options.recras_hostname)
+        ) {
+            throw new Error(
+                this.languageHelper.translate('ERR_INVALID_HOSTNAME')
+            );
         }
         if (options.redirect_url) {
             if (options.redirect_url.indexOf('http://') === -1 && options.redirect_url.indexOf('https://') === -1) {