+++ changelog.md
@@ -3,6 +3,7 @@
 ## master
 * Disable date selection if there are min/max amount or dependency errors
 * Fix "NaN" price when booking size input field was cleared
+* Add `previewTimes` option to disable programme times preview for online bookings
 
 ## 0.12.0 (2018-11-30)
 * Add `autoScroll` option to online bookings