+++ changelog.md
@@ -4,6 +4,13 @@
 * Disable date selection if there are min/max amount or dependency errors
 * Fix "NaN" price when booking size input field was cleared
 * Add `previewTimes` option to disable programme times preview for online bookings
+* Add `recras-completed` and `recras-active` classes. `recras-completed` is added to each section
+ the user has successfully completed, `recras-active` is added to the currently active section.
+  
+ **Note for developers** when returning to an already completed section (i.e. changing amounts 
+ after selecting a date and time) can make multiple sections active. Entering an invalid amount 
+ to the already completed amounts section will leave the section completed. If you want to do 
+ things like hide completed sections, be aware of this and test thoroughly!
 
 ## 0.12.0 (2018-11-30)
 * Add `autoScroll` option to online bookings