+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.13.1 (2019-01-07)
+* fix the test for development instances
+
 ## 0.13.0 (2019-01-07)
 * Disable date selection if there are min/max amount or dependency errors
 * Fix "NaN" price when booking size input field was cleared